requirements.txt
pandas
numpy
scikit-learn
spotipy
streamlit
matplotlib
seaborn
plotly
python-dotenv